feat: add password checker and setup sunshine creds

+++ fishtank/run.ps1
@@ -17,6 +17,12 @@ $set_password = yn_prompt "Set a password for this admin account?"
 
 if ($set_password) {
     $pw = Read-Host -AsSecureString -Prompt "Enter a password"
+    $pw2 = Read-Host -AsSecureString -Prompt "Re-enter the password"
+
+    if ([System.Net.NetworkCredential]::new("", $pw).Password -ne [System.Net.NetworkCredential]::new("", $pw2).Password) {
+        Write-Output "Passwords do not match. Exiting..."
+        exit
+    }
 
     # Set the current users password to the entered password
     $UserAccount = [System.Security.Principal.WindowsIdentity]::GetCurrent().Name
@@ -25,9 +31,19 @@ if ($set_password) {
     Write-Output "Password set successfully."
 }
 
+$install_sunshine = yn_prompt "Install sunshine software for remote desktop?"
+
+if ($install_sunshine) {
+    $sunshine_uname = Read-Host -Prompt "Enter a username for the sunshine account"
+    $sunshine_password = Read-Host -AsSecureString -Prompt "Enter a password for the sunshine account"
+    $sunshine_password_clean = [System.Net.NetworkCredential]::new("", $sunshine_password).Password
+}
+
 $install_common = yn_prompt "Install common software?"
 $install_dev = yn_prompt "Install developer software?"
 $install_gaming_room = yn_prompt "Install gaming room software? (Adds a user, ollama, wallpaper, and maps network drives)"
+$install_scripts = yn_prompt "Install scripts? (Debloat, auto-update, etc.)"
+$logout_after = yn_prompt "Log out after installation?"
 
 Unblock-File choco-installer.ps1
 . .\choco-installer.ps1
@@ -96,6 +112,15 @@ function execute_scripts_recursive($directory) {
     }
 }
 
+if ($install_sunshine) {
+    choco install -y sunshine
+    $sunshine_binary = "C:\Program Files\Sunshine\sunshine.exe"
+    $sunshine_creds = "--creds `"$sunshine_uname`" `"$sunshine_password_clean`""
+    Start-Process -FilePath $sunshine_binary -ArgumentList $sunshine_creds -WindowStyle Hidden -Wait
+    Stop-Process -Name "sunshine" -Force -ErrorAction SilentlyContinue
+    Start-Process -FilePath $sunshine_binary -ArgumentList $sunshine_creds -WindowStyle Hidden
+}
+
 if ($install_common) {
     install_choco_packages_recursive "apps/common"
     execute_scripts_recursive "scripts/common"
@@ -111,11 +136,15 @@ if ($install_gaming_room) {
     execute_scripts_recursive "scripts/gaming-room"
 }
 
-execute_scripts_recursive "scripts"
+if ($install_scripts) {
+    execute_scripts_recursive "scripts"
+}
 
 Write-Output "Fishtank is set up!"
 
-# Log out in 5 seconds
-Write-Output "Logging out in 5 seconds..."
-Start-Sleep -Seconds 5
-shutdown /l
+if ($logout_after) {
+    # Log out in 5 seconds
+    Write-Output "Logging out in 5 seconds..."
+    Start-Sleep -Seconds 5
+    shutdown /l
+}
